Fix all linter errors
