+++ app/panel-android/components/content/FixedMenu.jsx
@@ -1,74 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-
-class MenuItem extends React.Component {
-	constructor(props) {
-		super(props);
-
-		this.state = {
-			opening: false,
-		};
-	}
-
-	menuItemClicked = () => {
-		this.setState({
-			opening: true,
-		});
-
-		this.props.updateHeadeText(this.props.title);
-	}
-
-	closeButtonClicked = () => {
-		this.setState({
-			opening: false,
-		});
-
-		this.props.updateHeadeText('');
-	}
-
-	switcherClicked = () => {
-		this.context.callGlobalAction({
-			actionName: 'cliqzFeatureToggle',
-			actionData: {
-				currentState: this.props.active,
-				type: this.props.type,
-			},
-		});
-	}
-
-	render() {
-
-		return (
-			<div className="menuItemWrapper">
-				<div onClick={this.menuItemClicked} className="menuItemOverview">
-					<span className={this.props.type}>{this.props.numData}</span>
-					<span className="title">{this.props.title}</span>
-					<p className="description">{this.props.description}</p>
-				</div>
-				<span onClick={this.switcherClicked} className={`switcher ${this.props.active ? 'active' : ''}`}></span>
-				<div className={`menuItemContent ${this.state.opening ? 'opening' : ''}`}>
-					<span className={this.props.type}>{this.props.numData}</span>
-					<p className="headline">{this.props.headline}</p>
-					<p className="description">{this.props.description}</p>
-					<button onClick={this.closeButtonClicked} className="close"></button>
-				</div>
-			</div>
-		);
-	}
-}
-
-MenuItem.propTypes = {
-	active: PropTypes.bool,
-	type: PropTypes.string,
-	title: PropTypes.string,
-	numData: PropTypes.number,
-	headline: PropTypes.string,
-	description: PropTypes.string,
-};
-
-MenuItem.contextTypes = {
-	callGlobalAction: PropTypes.func,
-};
+import MenuItem from './MenuItem';
 
 export default class FixedMenu extends React.Component {
 	constructor(props) {
@@ -76,24 +8,7 @@ export default class FixedMenu extends React.Component {
 		this.state = {
 			open: false,
 			currentMenuItemText: this.defaultHeaderText,
-		}
-	}
-
-	toggleMenu = () => {
-		const currentState = this.state.open;
-		this.setState({
-			open: !currentState,
-		})
-	}
-
-	updateHeadeText = (text) => {
-		if (!text) {
-			text = this.defaultHeaderText;
-		}
-
-		this.setState({
-			currentMenuItemText: text,
-		});
+		};
 	}
 
 	get defaultHeaderText() {
@@ -117,7 +32,6 @@ export default class FixedMenu extends React.Component {
 	}
 
 	getCount = (type) => {
-		console.log('this.props.panel', this.props.panel);
 		let total = 0;
 		switch (type) {
 			case 'enable_anti_tracking':
@@ -146,8 +60,22 @@ export default class FixedMenu extends React.Component {
 		}
 	}
 
-	render() {
+	toggleMenu = () => {
+		const currentState = this.state.open;
+		this.setState({
+			open: !currentState,
+		});
+	}
 
+	updateHeadeText = (text) => {
+		const textToShow = text || this.defaultHeaderText;
+
+		this.setState({
+			currentMenuItemText: textToShow,
+		});
+	}
+
+	render() {
 		return (
 			<div className={`fixed-menu ${this.state.open ? 'opened' : ''}`}>
 				<div onClick={this.toggleMenu} className="menuHeader">
@@ -158,33 +86,33 @@ export default class FixedMenu extends React.Component {
 						<MenuItem
 							active={this.props.panel.enable_anti_tracking}
 							updateHeadeText={this.updateHeadeText}
-							type={'anti_tracking'}
-							title={'Enhanced Anti-Tracking'}
+							type="anti_tracking"
+							title="Enhanced Anti-Tracking"
 							numData={this.getCount('enable_anti_tracking')}
-							headline={'Personal data points anonymized'}
-							description={'Anonymize unblocked and unknown trackers for greater browsing protection.'}
+							headline="Personal data points anonymized"
+							description="Anonymize unblocked and unknown trackers for greater browsing protection."
 						/>
 					</li>
 					<li className="menuItem">
 						<MenuItem
 							active={this.props.panel.enable_ad_block}
 							updateHeadeText={this.updateHeadeText}
-							type={'ad_block'}
-							title={'Enhanced Ad Blocking'}
+							type="ad_block"
+							title="Enhanced Ad Blocking"
 							numData={this.getCount('enable_ad_block')}
-							headline={'Advertisements blocked'}
-							description={'Block all advertisements on websites you visit.'}
+							headline="Advertisements blocked"
+							description="Block all advertisements on websites you visit."
 						/>
 					</li>
 					<li className="menuItem">
 						<MenuItem
 							active={this.props.panel.enable_smart_block}
 							updateHeadeText={this.updateHeadeText}
-							type={'smart_block'}
-							title={'Smart Blocking'}
+							type="smart_block"
+							title="Smart Blocking"
 							numData={this.getCount('enable_smart_block')}
-							headline={'Smart Blocking'}
-							description={'Automatically block and unblock trackers to optimize page performance.'}
+							headline="Smart Blocking"
+							description="Automatically block and unblock trackers to optimize page performance."
 						/>
 					</li>
 				</ul>
@@ -194,6 +122,11 @@ export default class FixedMenu extends React.Component {
 }
 
 FixedMenu.propTypes = {
-	panel: PropTypes.object,
-	cliqzModuleData: PropTypes.object,
+	panel: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+	cliqzModuleData: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+};
+
+FixedMenu.defaultProps = {
+	panel: {},
+	cliqzModuleData: {},
 };
