+++ app/panel-android/components/content/Tabs.jsx
@@ -1,36 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import Tab from './Tab';
 
-export class Tab extends React.Component {
-
-	handleTabClick = (event) => {
-		event.preventDefault();
-		this.props.onClick(this.props.tabIndex);
-	}
-
-	render() {
-		return (
-			<li className="tab-item">
-				<a
-					className={`tab-link ${this.props.linkClassName} ${this.props.isActive ? 'active' : ''}`}
-					onClick={this.handleTabClick}
-				>
-					{this.props.tabLabel}
-				</a>
-			</li>
-		);
-	}
-}
-
-Tab.propTypes = {
-	onClick: PropTypes.func,
-	tabIndex: PropTypes.number,
-	isActive: PropTypes.bool,
-	tabLabel: PropTypes.string.isRequired,
-	linkClassName: PropTypes.string.isRequired
-};
-
-export class Tabs extends React.Component {
+export default class Tabs extends React.Component {
 	constructor(props) {
 		super(props);
 
@@ -65,6 +37,7 @@ export class Tabs extends React.Component {
 		if (children[activeTabIndex]) {
 			return children[activeTabIndex].props.children;
 		}
+		return null;
 	}
 
 	render() {
