+++ app/scss/android/content/_accordions.scss
@@ -75,9 +75,6 @@
 				text-transform: uppercase;
 
 				& > span {
-					&.total-trackers {
-						// color: #aaa;
-					}
 
 					&.blocked-trackers {
 						margin-left: 5px;
